# Conflicts:
#	CHANGELOG.md

+++ src/network.h
@@ -123,7 +123,7 @@ T Network::udpBlockingReceiveUnsafe() try {
   int bytes_received =
       udp_socket_.receiveFrom(buffer.data(), static_cast<int>(buffer.size()), udp_server_address_);
 
-  if (bytes_received != buffer.size()) {
+  if (bytes_received != static_cast<int>(buffer.size())) {
     throw ProtocolException("libfranka: incorrect object size");
   }
